ISSITES-5639 Sync article folders
