+++ config.json
@@ -7,7 +7,7 @@
 	"API_PAGESIZE": "300",
 	"WEBDAV_IMAGE_URL_PREFIX": "/webdav/learn-python-import/document_library/",
 	"DOCUMENT_IMPORT_LIMIT": 1,
-	"ARTICLE_IMPORT_LIMIT": 1,
+	"STRUCTURED_CONTENT_IMPORT_LIMIT": 1,
 	"API_PAGE_LIMIT": -1,
 	"DOCUMENT_TYPE": "Learn Synced File",
 	"GITHUB_EDIT_LINK_BASE_URL" : "https://github.com/liferay/liferay-learn/edit/master/docs/"
