+++ get_articles.py
@@ -1,5 +1,6 @@
 from configuration import config
-import get_all_items
+from get_all_items import get_all_items
+from liferay_api.structured_content_rest import get_site_structured_contents_page
 import json
 import logging
 import oauth_token
@@ -8,41 +9,43 @@ import util
 from jsonpath_ng.ext import parse
 
 
-def get_articles():
+def get_liferay_site_structured_contents_by_friendlyurlpath():
     sha_256sum_jsonpath = parse(
         '$.contentFields[?(@.label=="sha_256sum")].contentFieldValue.data'
     )
-    articles = get_all_items.get_all_items(get_article_batch)
+    liferay_site_structured_contents = get_all_items(get_site_structured_contents_page)
+    util.save_as_json(
+        "liferay_site_structured_contents", liferay_site_structured_contents
+    )
 
-    articles_by_friendlyurlpath = {}
-    for article in articles:
-        sha_256sum_jsonpath_find_result = sha_256sum_jsonpath.find(article)
+    # Only consider articles that are of the Learn Sync structure
+    liferay_site_learn_structured_contents = filter(
+        lambda site_structured_content: site_structured_content["contentStructureId"]
+        == config["ARTICLE_STRUCTURE_ID"],
+        liferay_site_structured_contents,
+    )
+
+    liferay_site_structured_contents_by_friendlyurlpath = {}
+    for liferay_site_structured_content in liferay_site_learn_structured_contents:
+        sha_256sum_jsonpath_find_result = sha_256sum_jsonpath.find(
+            liferay_site_structured_content
+        )
         sha_256sum = (
             sha_256sum_jsonpath_find_result[0].value
             if len(sha_256sum_jsonpath_find_result) == 1
             else ""
         )
 
-        articles_by_friendlyurlpath[article["friendlyUrlPath"]] = {
-            "id": article["id"],
-            "articleId": article["key"],
+        liferay_site_structured_contents_by_friendlyurlpath[
+            liferay_site_structured_content["friendlyUrlPath"]
+        ] = {
+            "id": liferay_site_structured_content["id"],
+            "articleId": liferay_site_structured_content["key"],
             "sha_256sum": sha_256sum,
         }
 
-    util.save_as_json("articles_by_friendlyurlpath", articles_by_friendlyurlpath)
-    return articles_by_friendlyurlpath
-
-
-@oauth_token.api_call(200)
-def get_article_batch(page):
-    logger = logging.getLogger(__name__)
-    headers = {
-        "Accept": "application/json",
-        "Authorization": oauth_token.authorization,
-        "Content-Type": "application/json",
-    }
-
-    get_uri = f"{config['OAUTH_HOST']}/o/headless-delivery/v1.0/sites/{config['SITE_ID']}/structured-contents?fields=key,contentFields,id,friendlyUrlPath&page={page}&pageSize={config['API_PAGESIZE']}"
-
-    logger.info(f"Fetching article page {page}")
-    return requests.get(get_uri, headers=headers)
+    util.save_as_json(
+        "liferay_site_structured_contents_by_friendlyurlpath",
+        liferay_site_structured_contents_by_friendlyurlpath,
+    )
+    return liferay_site_structured_contents_by_friendlyurlpath
